Introduce `#[DomainCommandHandler]` attribute for use cases
With this new attribute, a use case can directly bind a method to a command,
instead of creating a separate command + handler (the classic command bus way).

Since the classic command handler often does not do more than (a) get use case,
(b), call action and (c) save use case, the command handler dus not add much
extra value. Therefore, being able to bind the action method in the use case
to the command itself, saves the need for a command handler.

Note: this is optional, when needed, a command handler can still be created.