Introduce generic `UseCaseCommandHandler`
This is the command handler which will handle all incoming commands, which
are bind to a `#[DomainCommandHandler]` method in a use case.

Based on creation policy it will automatically create a use case if not present
yet.

Note: In order to make this work, the registering of command to this generic
command handler needs to be done custom, depending on the command bus vendor
which is being used.