Refactor `AttributeResolver:getMethodsWithAttribute`
To return a list of array{ReflectionMethod, Attribute}, so that any resolver
can do what it needs. With this change, the generic AttributeResolver gets
less responsibility, and just retrieves methods without knowing why.

The cached decorator will not work (easy) with attribute instances nor
reflection objects, but that will be solved in a next iteration with cached
resolver decorators.