Improve Sagas: Allow Sagas linked to multiple SagaIds
In this way, it is easier to create Sagas which listen to several SagaIds.

Often a Saga subscribes to events which have different identifiers.
When a Saga can only link to one SagaId, the options for a Saga are limited.
With this change, a Saga can subscribe to any event, as long as it is linked
to any SagaId defined in the event.